pyproject.toml
[project]
name = "fordera"
version = "0.1.0"
description = "Ford F-series front profile classifier and dichotomous key generator"
requires-python = ">=3.9"
dependencies = [
    "torch",
    "torchvision",
    "Pillow",
    "easyocr",
    "scikit-learn",
    "numpy",
    "matplotlib",
    "graphviz",
    "marimo",
    "opencv-python-headless",
]

[project.optional-dependencies]
dev = ["pytest"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
